JsonResponseObjectConsumer to attempt parsing JSON content in case of a 2xx response status only (fixes #4)
